Fix client issue
Thank you for that simple and functional library! 
Got one issue: when explicitly declaring 'userTokenInterceptor' and 'userLoggerInterceptor' and making some default request (say, auth), 'non-null' assertions related to these fields fail and the error goes up to the widget tree, so I propose to assign default values in the constructor.